backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
flask==2.3.2
flask-cors==3.0.10
numpy==1.24.3
pillow==9.5.0
scikit-learn==1.2.2
opencv-python-headless==4.7.0.72
torch==2.0.1
torchvision==0.15.2
transformers==4.29.2
timm==0.9.2
kornia==0.6.12
einops==0.6.1
reportlab==4.0.0
pyembroidery==1.4.11
requests==2.31.0
gunicorn==20.1.0
scikit-image==0.21.0
